Adding option to configure services(CP/DP) and their huge page size.
In install.sh a new option(no. 2) is added to configure services to be
build. On selecting option 2, user will be prompted to a submenu to
select service as:
1. CP Only
2. DP Only
3. Collocated CP and DP.
For the selected service, user will be displayed default memory size, with
option to edit the memory size.
Install script will internally calculate number of huge pages and update
in to /etc/sysctl.conf file.

Based on the selected service, list of options in the 'Download and Install'
section will change.

'Build NGIC' option will build only the selected service and its dependencies.

Change-Id: Icd08a57e5292469a98e28d57a435fe51f9e73e23